pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent_pipeline_framework"
# Package name will be agent_pipeline_framework
version = "0.1.0"
description = "A robust, generic agent pipeline framework for running sequential agent workflows"
readme = "README.md"
authors = [{name = "Kilo", email = "kilo@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "loguru>=0.7.0",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "docs*"]